Use legacy context manager
